refactor(arena/inspect): split renderer into per-section files (#1456)
